Add files via upload
The Sub-saharan rerion is worst affected by access to clean water according to the UN. The region Lags way behind. Majority of the people in Sub-saharan Africa live in Rural areas. Additionally, most of the population have broken taps in their homes, they hence rely on shared taps. The most popular water source of water therefore is wells and shared taps. Very few rely on rivers as the region is drier. As a result queue times are often long. Another key finding is that the crime is proportionate to to the queue times. majority of the affected people are women with harassments as the major crime against them. This is true as women are the major water collectors and thus are left vulnerable to crimes!